GEN6-912: Initial simply Redis deployment
Redis will begin to be used as part of Gen6 as a persistent data store.

This commit adds a standalone master instance of Redis as a stateful set
and exposes it via a service. Redis is configured to use AOF persistence
writing out each second. The persistence files are stored onto a
persistent volume to withstand restarts without data loss.