Add FusionAuth base config (#59)
INTENT
Add Fusion Auth base for usage by Kustomize in projects

DETAIL
Added Fusion Auth. This is based on their published works, but tweaked so that it actually works. It also has DB integration which relies on Crunch Data operator. This may, or may not be what we want, but I think an opinionated config is a good config.