Remove Postgres from Fusion (#73)
INTENT
Remove Postgres from the FusionAuth base

DETAIL
Currently Postgres DB's are setup via the operator in the FusionAuth base. this is a bit of an anti-pattern as it ties both together, and also, not what the rest of the apps do. This removes the postgresDB, which can be provided when over-riding in the customer repos.