Remove DB from Fusion (#82)
INTENT
Decouple the FusionAuth struct from the DB

DETAIL
The fusionauth struct is tied to an operator supplied DB. This both limits its ability to be deployed in certain environments, and has a tendency to remove data when upgrading etc. This PR removes the DB, and it will be re-added into the project structs where appropriate.