FIX: Set explicit Kubeval API version (#147)
Kubeval fails silently if it is not able to download schemas that match
the Kubernetes API Version passed in with the `--kubernetes-version/-v`
flag. This sets an explicit Kubernetes version to validate the schemas
against. The value of `KUBERNETES_VERSION` that was in use before was
prefixed with a v that was not a valid version in the
instrumenta/kubernetes-schema-json repo.